* Converted c-style casts to c++ casts (except dynamic_cast) where possible. The point here was to make it really easy to find the often non-portable reinterpret_casts to integer types and back, and make it easier to find and evaluate many other casts. * In many places casts were not necessary and were eliminated. * to_string and from_string moved out of SoarKernel's misc.h and in to shared project so that other projects could use it that didn't depend on the kernel. Removed kernel's misc.h and put the rest of its contents in utilities.cpp/h * Fixed an issue with Float and Int element (ClientSML) where const methods were actually const-casting away their own 'this' pointer and modifying themselves. * Changed some memory functions to use size_t when possible to make the code more portable * Cleaned up some of the rete's inline functions that were initially created by copy-and-pasting the old macros and therefore had many extra parens required in macros but not inline functions. The result is much easier to look at and understand. * All tests pass and things work fine on windows, testing unix/mac after this commit.

+++ ClientSML/include/sml_ClientKernel.h
@@ -493,7 +493,7 @@ public:
 	* @returns The result of executing the run command.
 	*		   The output from during the run is sent to a different callback.
 	*************************************************************/
-	char const* RunAllAgents(unsigned long numberSteps, smlRunStepSize stepSize = sml_DECISION, smlRunStepSize interleaveStepSize = sml_PHASE) ;
+	char const* RunAllAgents(int numberSteps, smlRunStepSize stepSize = sml_DECISION, smlRunStepSize interleaveStepSize = sml_PHASE) ;
 	char const* RunAllAgentsForever(smlRunStepSize interleaveStepSize = sml_PHASE) ;
 
 	/*************************************************************
